core: Inline AbstractServerImplBuilder
